+++ frs/include/huaweicloud/frs/v2/FrsRegion.h
@@ -22,6 +22,7 @@ public:
 
     static std::map<std::string, Region> initRegions() {
         std::map<std::string, Region> regions;
+        regions["cn-north-1"] = Region("cn-north-1","https://face.cn-north-1.myhuaweicloud.com") ;
         regions["cn-north-4"] = Region("cn-north-4","https://face.cn-north-4.myhuaweicloud.com") ;
         regions["cn-south-1"] = Region("cn-south-1","https://face.cn-south-1.myhuaweicloud.com") ;
         regions["cn-east-3"] = Region("cn-east-3","https://face.cn-east-3.myhuaweicloud.com") ;