+++ ims/include/huaweicloud/ims/v2/model/ShowVersionRequest.h
@@ -1,0 +1,72 @@
+
+#ifndef HUAWEICLOUD_SDK_IMS_V2_MODEL_ShowVersionRequest_H_
+#define HUAWEICLOUD_SDK_IMS_V2_MODEL_ShowVersionRequest_H_
+
+#include <huaweicloud/ims/v2/ImsExport.h>
+
+#include <huaweicloud/core/utils/ModelBase.h>
+#include <huaweicloud/core/http/HttpResponse.h>
+
+#include <string>
+
+namespace HuaweiCloud {
+namespace Sdk {
+namespace Ims {
+namespace V2 {
+namespace Model {
+
+using namespace HuaweiCloud::Sdk::Core::Utils;
+using namespace HuaweiCloud::Sdk::Core::Http;
+
+/// <summary>
+/// Request Object
+/// </summary>
+class HUAWEICLOUD_IMS_V2_EXPORT  ShowVersionRequest
+    : public ModelBase
+{
+public:
+    ShowVersionRequest();
+    virtual ~ShowVersionRequest();
+
+    /////////////////////////////////////////////
+    /// ModelBase overrides
+
+    void validate() override;
+
+    web::json::value toJson() const override;
+    bool fromJson(const web::json::value& json) override;
+
+    /////////////////////////////////////////////
+    /// ShowVersionRequest members
+
+    /// <summary>
+    /// API版本号。例如：v2.0
+    /// </summary>
+
+    std::string getVersion() const;
+    bool versionIsSet() const;
+    void unsetversion();
+    void setVersion(const std::string& value);
+
+
+protected:
+    std::string version_;
+    bool versionIsSet_;
+
+#ifdef RTTR_FLAG
+    RTTR_ENABLE()
+public:
+    ShowVersionRequest& dereference_from_shared_ptr(std::shared_ptr<ShowVersionRequest> ptr) {
+        return *ptr;
+    }
+#endif
+};
+
+
+}
+}
+}
+}
+}
+
+#endif // HUAWEICLOUD_SDK_IMS_V2_MODEL_ShowVersionRequest_H_