+++ ocr/include/huaweicloud/ocr/v1/model/IdCardRequestBody.h
@@ -75,6 +75,15 @@ public:
     void unsetreturnVerification();
     void setReturnVerification(bool value);
 
+    /// <summary>
+    /// 识别到的文字块的区域位置信息。可选值包括：  - true：返回各个文字块区域  - false：不返回各个文字块区域  如果无该参数，系统默认不返回文字块区域。如果输入参数不是Boolean类型，则会报非法参数错误。 
+    /// </summary>
+
+    bool isReturnTextLocation() const;
+    bool returnTextLocationIsSet() const;
+    void unsetreturnTextLocation();
+    void setReturnTextLocation(bool value);
+
 
 protected:
     std::string image_;
@@ -85,6 +94,8 @@ protected:
     bool sideIsSet_;
     bool returnVerification_;
     bool returnVerificationIsSet_;
+    bool returnTextLocation_;
+    bool returnTextLocationIsSet_;
 
 };