+++ ocr/include/huaweicloud/ocr/v1/model/VatInvoiceResult.h
@@ -312,6 +312,15 @@ public:
     void unsetconfidence();
     void setConfidence(const Object& value);
 
+    /// <summary>
+    /// 文本框在原图位置。输出左上、右上、右下、左下四个点坐标。当“return_text_location”设置为“true”时才返回。 
+    /// </summary>
+
+    Object getTextLocation() const;
+    bool textLocationIsSet() const;
+    void unsettextLocation();
+    void setTextLocation(const Object& value);
+
 
 protected:
     std::string type_;
@@ -374,6 +383,8 @@ protected:
     bool itemListIsSet_;
     Object confidence_;
     bool confidenceIsSet_;
+    Object textLocation_;
+    bool textLocationIsSet_;
 
 };